OpenApi PayDocumentationModifier for Sylius >=1.13 and Swagger PayDocumentationNormalizer for Sylius <1.12
